extract entitlements module for background js
